Add Alert

+++ src/hooks/battery/index.jsx
@@ -1,5 +1,5 @@
 import React, { createContext , useState, useEffect, useContext, useCallback} from 'react';
-// import Swal from 'sweetalert2';
+import Swal from 'sweetalert2';
 import {api} from '../../services/api';
 import { useServicesBattery } from '../../hooks/servicesBattery';
 import { useService } from '../../hooks/service';
@@ -10,6 +10,22 @@ const BatteryContext = createContext();
 const BatteryProvider = ({children}) => {
   const [batteries, setBatteries] = useState([]);
   
+  const alert = async () =>{
+    try {
+       const alert =  await Swal.fire({
+            title: 'Você tem certeza?',
+            text: `Você não poderá reverter essa ação!.`,
+            icon: 'warning',
+            showCancelButton: true,
+            confirmButtonText: 'Sim!',
+            cancelButtonText: 'Não, não!'
+        });
+       return !!(alert.value && alert.value === true);
+    } catch (error) {
+        console.log('error:', error);
+        return false;
+    }
+  }
 
   const history = useHistory();
 
@@ -27,41 +43,16 @@ const BatteryProvider = ({children}) => {
   useEffect(()=>{
     loadBatteries();
   },[loadBatteries])
-
-
-  
-
+ 
   const handleDelete = async (id) => {
-    // Swal.fire({
-    //   icon: 'warning',
-    //   title: 'Você tem certeza?',
-    //   text: "Você não poderá reverter essa ação!",
-    //   showCancelButton: true,
-    //   confirmButtonText: 'Sim, delete!',
-    //   cancelButtonText: 'Não, não delete!'
-    // }).then(result => {
-    //   if (result.value) {
-    //     Swal.fire({
-    //       icon: 'success',
-    //       title: 'Deletado!',
-    //       text: `Deletado com sucesso.`,
-    //       showConfirmButton: false,
-    //       timer: 1500
-    //     });
+  
+    if(await alert()){
+      await api.delete(`batteries/${id}`);
+          
+      setBatteries(batteries.filter(battery => battery.id !== id));
 
-    //     api.delete(`/batteries/${id}`);
-        
-    //     setBatteries(batteries.filter(battery => battery.id !== id));
-
-    //     history.push('/');
-    //   }
-    // });
-    
-    await api.delete(`batteries/${id}`);
-        
-    setBatteries(batteries.filter(battery => battery.id !== id));
-
-    history.push('/');
+      history.push('/');
+    }
   };
   async function handleChargeBattery (id) {
     // Swal.fire({
@@ -97,105 +88,37 @@ const BatteryProvider = ({children}) => {
     //   }
     // });
 
-      // Swal.fire({
-    //   icon: 'warning',
-    //   title: 'Você tem certeza?',
-    //   text: "Você não poderá reverter essa ação!",
-    //   showCancelButton: true,
-    //   confirmButtonText: 'Sim!',
-    //   cancelButtonText: 'Não, não!'
-    // }).then(result => {
-    //   if (result.value) {
-    //     Swal.fire({
-    //       icon: 'success',
-    //       title: 'Carregando!',
-    //       text: `Ação realizada com sucesso.`,
-    //       showConfirmButton: false,
-    //       timer: 1500
-    //     });
-
-
-    //      api.post(`/batteries-services/${id}`).then(
-    //       batteryUpdated => {
-    //         const batteriesUpdated = batteries.map(battery =>
-    //           battery.id !== batteryUpdated.data.id ? battery : batteryUpdated.data,
-    //         );
+ 
+    if(await alert()){
+      const batteryUpdated = await api.post(`batteries-services/${id}`)
+    
+      const batteriesUpdated = batteries.map(battery =>
+        battery.id !== batteryUpdated.data.id ? battery : batteryUpdated.data,
+      );
+      setBatteries([...batteriesUpdated]);
       
-    //         setBatteries([...batteriesUpdated]);
-    //       }
-    //     );
-    //     loadServicesBatteries();
-    //     loadServices();
-    //     history.push('/');
-    //   }
-    // });
-    api.post(`batteries-services/${id}`).then(
-      batteryUpdated => {
-        const batteriesUpdated = batteries.map(battery =>
-          battery.id !== batteryUpdated.data.id ? battery : batteryUpdated.data,
-        );
-  
-        setBatteries([...batteriesUpdated]);
-      }
-    );
-    loadServicesBatteries();
-    loadServices();
-    history.push('/');
+      loadServicesBatteries();
+      loadServices();
+      history.push("/batteries/dashboard");
+    }
   };
   async function handleFinishBatteryCharge (id) {
-    // Swal.fire({
-    //   icon: 'warning',
-    //   title: 'Você tem certeza?',
-    //   text: "Você não poderá reverter essa ação!",
-    //   showCancelButton: true,
-    //   confirmButtonText: 'Sim!',
-    //   cancelButtonText: 'Não, não!'
-    // }).then(result => {
-    //   if (result.value) {
-    //     Swal.fire({
-    //       icon: 'success',
-    //       title: 'Carregando!',
-    //       text: `Ação realizada com sucesso.`,
-    //       showConfirmButton: false,
-    //       timer: 1500
-    //     });
+  
+    if(await alert()){
+      
+      const batteryUpdated = await api.post(`batteries-full-charge/${id}`);
 
-
-    //      api.post(`/batteries-full-charge/${id}`).then(
-    //       batteryUpdated => {
-    //         const batteriesUpdated = batteries.map(battery =>
-    //           battery.id !== batteryUpdated.data.id ? battery : batteryUpdated.data,
-    //         );
-        
-    //         setBatteries([...batteriesUpdated]);
-    //       }
-    //     );
-
-    //     loadServicesBatteries();
-    //     loadServices();
-    //     history.push('/');
-    //   }
-    // });
-
-    
-    api.post(`batteries-full-charge/${id}`).then(
-      batteryUpdated => {
-        const batteriesUpdated = batteries.map(battery =>
-          battery.id !== batteryUpdated.data.id ? battery : batteryUpdated.data,
-        );
-    
-        setBatteries([...batteriesUpdated]);
-      }
-    );
-
-    loadServicesBatteries();
-    loadServices();
-    history.push('/');
+      const batteriesUpdated = batteries.map(battery =>
+        battery.id !== batteryUpdated.data.id ? battery : batteryUpdated.data,
+      );
+      
+      setBatteries([...batteriesUpdated]);
+      
+      loadServicesBatteries();
+      loadServices();
+      history.push('/');
+    }
   };
-
-
- 
-
 
   return (
     <BatteryContext.Provider value={{
